Rename files, fix links and add next steps
